Restore the amazon scraper
