+++ src/interfaces/exitoScraper.interface.go
@@ -1,0 +1,5 @@
+package interfaces
+
+type ExitoScraper struct {
+	Scraper
+}
