+++ src/services/cron.service.go
@@ -4,9 +4,7 @@ import (
 	"log"
 	"net/url"
 	"strings"
-	"webScraper/src/constants"
 	"webScraper/src/database"
-	"webScraper/src/interfaces"
 	"webScraper/src/services/scrapers"
 )
 
@@ -29,14 +27,9 @@ func CheckProducts() {
 		// Check the product based on the host
 		if strings.Contains(URL.Host, "amazon") {
 			// Make the Scraping to the page
-			// Create Object interface
-			scraper := interfaces.Scraper{
-				AllowedDomains: []string{constants.AMAZON_HALF_DOMAIN, constants.AMAZON_DOMAIN},
-			}
 
 			// Create Collector
-			collector := scraper.InitCollector()
-			scrapedProduct, err := scrapers.SendAmazonCollyRequest(URL.String(), scraper, collector)
+			scrapedProduct, err := scrapers.SendAmazonCollyRequest(URL.String())
 			if err != nil {
 				log.Println("Error Getting the data from the craping ", err)
 			}
